Move test function to tests
