+++ src/Microsoft.Management.Configuration.Processor/Extensions/ValueSetExtensions.cs
@@ -9,8 +9,6 @@ namespace Microsoft.Management.Configuration.Processor.Extensions
     using System;
     using System.Collections;
     using System.Collections.Generic;
-    using System.Text;
-    using System.Xml.Linq;
     using Windows.Foundation.Collections;
 
     /// <summary>
@@ -176,62 +174,6 @@ namespace Microsoft.Management.Configuration.Processor.Extensions
             }
 
             return true;
-        }
-
-        /// <summary>
-        /// Converts the value set to YAML like output.
-        /// </summary>
-        /// <param name="set">The set to output.</param>
-        /// <returns>The string.</returns>
-        public static string ToYaml(this ValueSet set)
-        {
-            StringBuilder sb = new StringBuilder();
-            ToYaml(set, sb);
-            return sb.ToString();
-        }
-
-        private static void ToYaml(ValueSet set, StringBuilder sb, int indentation = 0)
-        {
-            foreach (var keyValuePair in set)
-            {
-                bool addLine = true;
-
-                sb.Append(' ', indentation);
-                sb.Append(keyValuePair.Key);
-                sb.Append(": ");
-
-                if (keyValuePair.Value == null)
-                {
-                    sb.Append("null");
-                }
-                else
-                {
-                    switch (keyValuePair.Value)
-                    {
-                        case int i:
-                            sb.Append(i);
-                            break;
-                        case string s:
-                            sb.Append(s);
-                            break;
-                        case bool b:
-                            sb.Append(b);
-                            break;
-                        case ValueSet v:
-                            sb.AppendLine();
-                            ToYaml(v, sb, indentation + 2);
-                            addLine = false;
-                            break;
-                        default:
-                            throw new NotImplementedException($"Add ToYaml type `{keyValuePair.Value.GetType().Name}`");
-                    }
-                }
-
-                if (addLine)
-                {
-                    sb.AppendLine();
-                }
-            }
         }
     }
 }
